% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbox_danube_basin}
\alias{bbox_danube_basin}
\title{Polygon that includes the Danube River Basin}
\format{
\subsection{\code{bbox_danube_basin}}{

A simple feature collection with 1 feature and o fields
Geometry type: MULTIPOLYGON
Bounding box:  xmin: 8.1525 ymin: 42.08333 xmax: 29.73583 ymax: 50.245
Geodetic CRS:  WGS 84
}
}
\source{
QGIS
}
\usage{
bbox_danube_basin
}
\description{
A polygon of the Danube River Basin. Made with QGIS.
}
\keyword{datasets}
