% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coordinates.R
\name{check_coordinates}
\alias{check_coordinates}
\title{Check Coordinates for WGS84 and Decimal Degrees}
\usage{
check_coordinates(
  df,
  lat_col,
  lon_col,
  convert_to_numeric = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A data frame containing the coordinates.}

\item{lat_col}{The name of the column containing latitude values.}

\item{lon_col}{The name of the column containing longitude values.}

\item{convert_to_numeric}{Logical. If \code{TRUE}, attempts to convert the
coordinates to numeric. Default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, prints detailed information about invalid
and non-numeric coordinates. Default is \code{FALSE}.}
}
\value{
A list containing:
\itemize{
\item \code{invalid_latitudes}: Indices of latitude values that are out of range or invalid.
\item \code{invalid_longitudes}: Indices of longitude values that are out of range or invalid.
\item \code{non_numeric_latitudes}: Indices of non-numeric latitude values (if any).
\item \code{non_numeric_longitudes}: Indices of non-numeric longitude values (if any).
}
}
\description{
This function checks if the coordinates in a data frame are in the WGS84
format and in decimal degrees. It ensures that the latitude values are
between -90 and 90, and the longitude values are between -180 and 180.
If the \code{convert_to_numeric} parameter is \code{TRUE}, it attempts to convert
non-numeric coordinates to numeric. It also prints detailed information
if \code{verbose} is set to \code{TRUE}.
}
\examples{
df <- data.frame(latitude = c("45.5", "91", NA, "invalid", "12.5"),
                 longitude = c("-123.5", "200", "text", "-100", "55"))
check_coordinates(df, "latitude", "longitude", convert_to_numeric = TRUE, verbose = TRUE)
}
