% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species_name.R
\name{check_species_name}
\alias{check_species_name}
\title{Check Species Names with Variable Accuracy}
\usage{
check_species_name(
  data,
  col_species_name,
  verbose = FALSE,
  target_accuracy = 90,
  accuracy_decrement = NULL,
  synonym = FALSE
)
}
\arguments{
\item{data}{A data frame containing species names.}

\item{col_species_name}{A character string specifying the column in \code{data} with species names.}

\item{verbose}{Logical, if \code{TRUE}, additional information will be printed during execution.}

\item{target_accuracy}{Numeric, the starting accuracy (in percentage) for species name matching.}

\item{accuracy_decrement}{Numeric, optional. If provided, the function will loop through decreasing
accuracy levels, decrementing by this value.}

\item{synonym}{Logical, if \code{TRUE}, synonym names will be considered when checking species names.}
}
\value{
A data frame with species names checked against FishBase at the specified accuracy,
or a list of data frames if \code{accuracy_decrement} is provided.
}
\description{
This function checks the validity of species names in a data frame using the \code{check_names}
function from the \code{specleanr} package, allowing for flexible matching accuracy and synonym lookup.
It extends \code{check_names} by enabling loops through different matching accuracy levels if required.
}
\examples{
\dontrun{
  library(specleanr)
  data <- data.frame(species = c("Salmo salar", "Oncorhynchus mykiss"))
  result <- check_species_name(data, col_species_name = "species", target_accuracy = 95)
}
}
