% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_year_column.R
\name{check_year_column}
\alias{check_year_column}
\title{Check Missing Values and Validity of Year Column}
\usage{
check_year_column(
  df,
  col_year,
  year_range = c(1900, 2024),
  transform_numeric = TRUE
)
}
\arguments{
\item{df}{A data frame containing the data to be checked.}

\item{col_year}{The name of the column in the data frame that contains the year values.}

\item{year_range}{A numeric vector of length 2 specifying the valid range of years
(e.g., \code{c(1900, 2024)}).}

\item{transform_numeric}{Logical. If \code{TRUE}, attempts to convert non-numeric year
values to numeric. Defaults to \code{TRUE}.}
}
\value{
A list containing:
\itemize{
\item \code{missing_values}: Indices of missing values in the year column.
\item \code{invalid_years}: Indices of values that fall outside the valid year range.
}
}
\description{
This function checks for missing values in a specified column of a data frame
that contains years. It also ensures that the values are numeric and fall within
a valid range. If \code{transform_numeric} is \code{TRUE}, the function attempts to convert
non-numeric values to numeric.
}
\examples{
df <- data.frame(year = c("2001", "2005", NA, "two thousand and ten", "2018", "2050"))
check_year_column(df, "year", year_range = c(1900, 2024))
}
