% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gbif.R
\name{clean_gbif}
\alias{clean_gbif}
\title{Clean GBIF Occurrence Records}
\usage{
clean_gbif(
  df,
  coordinatePrecision = NULL,
  coordinateUncertaintyInMeters = NULL,
  buffer_centroid_countries = NULL,
  buffer_centroid_capitals = NULL,
  buffer_centroid_zoo_herbaria = NULL,
  decimalLongitude = "decimalLongitude",
  decimalLatitude = "decimalLatitude",
  speciesKey = "speciesKey",
  datasetKey = "datasetKey",
  year = "year",
  species = "species",
  duplicate = FALSE
)
}
\arguments{
\item{df}{A data frame containing species occurrence records,
typically downloaded using the \code{download_gbif_records()} function from this package.}

\item{coordinatePrecision}{Numeric. Remove records below this precision.}

\item{coordinateUncertaintyInMeters}{Numeric. Remove records below this uncertainty or specific common values (301, 3036, 999, 9999).}

\item{buffer_centroid_countries}{Numeric. Buffer size for removing records near country centroids.}

\item{buffer_centroid_capitals}{Numeric. Buffer size for removing records near capital centroids.}

\item{buffer_centroid_zoo_herbaria}{Numeric. Buffer size for removing records near zoo or herbaria centroids.}

\item{decimalLongitude}{String. Column name for longitude. Default is "decimalLongitude".}

\item{decimalLatitude}{String. Column name for latitude. Default is "decimalLatitude".}

\item{speciesKey}{String. Column name for species key. Default is "speciesKey".}

\item{datasetKey}{String. Column name for dataset key. Default is "datasetKey".}

\item{year}{String. Column name for the year of occurrence. Default is "year".}

\item{species}{String. Column name for the species. Default is "species".}

\item{duplicate}{Logical. Default is FALSE. Whether or not to remove duplicates that have
the same longitude, latitude, speciesKey, and datasetKey.}
}
\value{
A cleaned data frame with species occurrences.
}
\description{
This function cleans species occurrence records downloaded using the
\code{download_gbif_records()} function from this package. It applies filters based on
coordinate precision, uncertainty, and geographic buffers, removes records with low precision
or uncertainty, excludes records near country, capital, and zoo/herbaria centroids using methods
from the \code{CoordinateCleaner} package, and eliminates duplicates based on longitude,
latitude, speciesKey, and datasetKey. Additionally, it removes records
that do not contain species or year information.
}
\details{
The function uses the following methods from the \code{CoordinateCleaner} package:
\itemize{
\item \code{cc_cen()}: Removes records near country centroids.
\item \code{cc_cap()}: Removes records near capital centroids.
\item \code{cc_inst()}: Removes records near zoo or herbaria centroids.
}
}
\examples{
# Example usage
df <- download_gbif_records(...)
cleaned_df <- clean_gbif(df, coordinatePrecision = 0.01,
                         coordinateUncertaintyInMeters = 100,
                         buffer_centroid_countries = 10,
                         buffer_centroid_capitals = 10,
                         buffer_centroid_zoo_herbaria = 10,
                         decimalLongitude = "decimalLongitude",
                         decimalLatitude = "decimalLatitude",
                         speciesKey = "speciesKey",
                         datasetKey = "datasetKey",
                         year = "year",
                         species = "species")

}
