% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata_file.R
\name{create_metadata_file}
\alias{create_metadata_file}
\title{Create Metadata File}
\usage{
create_metadata_file(
  file_path,
  description,
  author,
  author_email = NULL,
  creation_date = Sys.Date(),
  source = NULL,
  licence = NULL,
  method = NULL,
  repo = NULL,
  doi = NULL,
  notes = NULL,
  metadata_folder = dirname(file_path)
)
}
\arguments{
\item{file_path}{Character. The full path of the file for which metadata is to be created.}

\item{description}{Character. A brief description of the file.}

\item{author}{Character. The name of the author or creator of the file.}

\item{author_email}{Character. The email of the author or creator of the file.}

\item{creation_date}{Date. The date when the file was created (default is the current date).}

\item{source}{Character. A URL or reference for the source of the data.}

\item{licence}{Character. The license under which the dataset is provided. Common licenses include:
\itemize{
\item \verb{CC0 1.0 (Public Domain Dedication)} - No rights reserved, free for any use.
\item \verb{CC BY 4.0 (Attribution)} - Free to use, with attribution.
\item \verb{CC BY-SA 4.0 (Attribution-ShareAlike)} - Free to use, with attribution and share-alike.
\item \verb{ODbL (Open Database License)} - Allows use, modification, and sharing, but derivatives must remain under ODbL.
\item \verb{CC BY-NC 4.0 (Attribution-NonCommercial)} - Non-commercial use only, with attribution.
\item \verb{CC BY-ND 4.0 (Attribution-NoDerivatives)} - You can share, but not modify the dataset.
\item \verb{GNU GPL} - Allows reuse and modification, but derivatives must also be under GPL.
\item \verb{MIT License} - Permissive, allows reuse in any context with attribution.
Specify the license and provide a link if applicable (e.g., \verb{CC BY 4.0 (https://creativecommons.org/licenses/by/4.0/)}).
}}

\item{method}{Character. A short description of actions conducted to obtain or modify the dataset (e.g., "crop", "filter") or an indication of the script used to obtain or modify the dataset.}

\item{repo}{Character. A URL or reference to the repository where the script used to obtain or modify the dataset can be found (e.g., GitHub, Zenodo).}

\item{doi}{Character. The Digital Object Identifier (DOI) for the dataset or related publication, if available.}

\item{notes}{Character. Any additional notes about the file.}

\item{metadata_folder}{Character. The folder where the generated metadata file will be saved.
If not specified, the metadata file will be saved in the same folder as the file.}

\item{version}{Character. The version of the file. If not provided, it will be automatically assigned.}
}
\value{
A metadata file saved in the specified metadata folder or the same folder as the file if \code{metadata_folder} is not provided.
}
\description{
This function generates a metadata file describing a specified file stored in a folder.
}
\examples{
\dontrun{
   create_metadata_file(file_path = "data/my_data.csv",
                        description = "Sample dataset",
                        author = "John Doe",
                        author_email = "john.doe@example.com",
                        creation_date = Sys.Date(),
                        source = "https://example.com",
                        licence = "CC BY 4.0",
                        method = "Filtered using script XYZ"L,
                        repo = "https://github.com/repo",
                        doi = "10.1234/zenodo.1234567",
                        notes = "No additional notes")
}
}
