% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gbif_records.R
\name{download_gbif_records}
\alias{download_gbif_records}
\title{Download GBIF Records for Specified Species within a Polygon}
\usage{
download_gbif_records(
  species_names,
  wkt,
  output_occur_path,
  gbif_user,
  gbif_pwd,
  gbif_email,
  import_to_r = TRUE
)
}
\arguments{
\item{species_names}{A vector of species names to download records for.}

\item{wkt}{The well-known text (WKT) representation of a polygon for spatial filtering.}

\item{output_occur_path}{The path where the downloaded records will be saved.}

\item{gbif_user}{Your GBIF username.}

\item{gbif_pwd}{Your GBIF password.}

\item{gbif_email}{Your GBIF email for notifications.}

\item{import_to_r}{A logical value indicating whether to import the downloaded data into R. Default is TRUE.}
}
\value{
A list containing the downloaded occurrence records (if \code{import_to_r} is TRUE)
and the citation for the GBIF download.
}
\description{
This function downloads occurrence records from GBIF for a given vector of species
names within a specified polygon, provided as a WKT string. The function sets up
GBIF credentials, checks for taxon keys, handles synonyms, and downloads records
based on the polygon and species names.
}
\details{
The function automatically filters records to ensure that occurrences with geospatial
issues or missing coordinates are excluded. Additionally, it excludes fossil and
living specimens from the download.
}
\examples{
# Define the path to the polygon file included in the package
polygon_path <- system.file("extdata", "danube_basin.gpkg", package = "yourPackageName")

# Convert the polygon to WKT format using the `get_wkt` function from your package
wkt <- get_wkt(bbox_danube_basin)

# Download occurrences for species of interest
output_occur_path <- "./downloaded_records"
gbif_data <- download_gbif_records(
  species_names = c("Hucho hucho", "Alburnoides bipunctatus", "Chondrostoma nasus"),
  wkt = wkt,
  output_occur_path = output_occur_path,
  gbif_user = "your_username",
  gbif_pwd = "your_password",
  gbif_email = "your_email@example.com",
  import_to_r = TRUE
)

# Access the downloaded records
occurrence_records <- gbif_data$raw_download

# Access the GBIF citation
gbif_citation <- gbif_data$gbif_download_citation

# Create a metadata file for the downloaded records
create_metadata_file(file_path = paste0(output_occur_path,"file_name.zip"),
description = "This file contains fish species occurrences from GBIF",
author = "Your Name",
author_email = "your_email@example.com",
source = gbif_citation)
}
