% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadm_danube_basin}
\alias{gadm_danube_basin}
\title{Danube River Basin Clipped from GADM Global Data}
\format{
\subsection{\code{gadm_danube_basin}}{

Simple feature collection with 13,724 features and 52 fields
Geometry type: MULTIPOLYGON
Dimension:     XY
Bounding box:  xmin: 8.1525 ymin: 42.08333 xmax: 29.73583 ymax: 50.245
Geodetic CRS:  WGS 84
}
}
\source{
Derived from GADM global administrative boundaries and clipped using the Danube River Basin extent in QGIS.
}
\usage{
gadm_danube_basin
}
\description{
This polygon represents the Danube River Basin, obtained by clipping global GADM administrative boundaries with the Danube River Basin extent. The dataset includes detailed administrative information for the region.
}
\keyword{datasets}
