% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wkt.R
\name{get_wkt}
\alias{get_wkt}
\title{Get WKT Representation of Polygon from sf Object in Counterclockwise Order}
\usage{
get_wkt(sf_obj)
}
\arguments{
\item{sf_obj}{An sf object containing a polygon geometry.}
}
\value{
A WKT string representing the polygon in counterclockwise winding order.
}
\description{
This function takes an sf polygon object, ensures that the polygon is in
counterclockwise winding order, and returns the WKT (Well-Known Text)
representation of the polygon.
}
\examples{
get_wkt(bbox_danube_basin)
}
