% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_metadata_app.R
\name{run_metadata_app}
\alias{run_metadata_app}
\title{Run the Metadata Generator Mini App (Dublin Core)}
\usage{
run_metadata_app()
}
\value{
Launches a miniUI Shiny app.
}
\description{
This function launches a miniUI Shiny app that allows the user to fill in metadata
according to Dublin Core standards and download an XML file.
}
