% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_points_on_map.R
\name{snap_points_on_map}
\alias{snap_points_on_map}
\title{Interactive Point Snapping on Map}
\usage{
snap_points_on_map(coords_df)
}
\arguments{
\item{coords_df}{A data frame containing columns \code{lat} and \code{lng} for latitude and longitude.}
}
\value{
A data frame with old and new coordinates and update status.
}
\description{
This function allows users to interactively snap points on a map by clicking
and dragging markers. It returns a data frame with old and new coordinates,
and indicates whether the coordinates were updated.
}
\examples{
coords <- data.frame(
id = 1:2,
lat = c(48.2082, 44.8176),   # Vienna, Belgrade
lng = c(16.3738, 20.4633))    # Vienna, Belgrade
updated_coords <- snap_points_on_map(coords)
}
