% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_join_gadm.R
\name{spatial_join_gadm}
\alias{spatial_join_gadm}
\title{Perform a Spatial Join of Points with GADM Administrative Boundaries}
\usage{
spatial_join_gadm(
  country_codes = c("AUT", "BIH", "BGR", "CHE", "CZE", "DEU", "HUN", "HRV", "MDA", "MNE",
    "POL", "ROU", "SRB", "SVK", "SVN", "UKR"),
  coords_df,
  lat_col = "decimalLatitude",
  lon_col = "decimalLongitude",
  sf_gadm = NULL,
  timeout = 600
)
}
\arguments{
\item{country_codes}{A character vector of ISO3 country codes for which GADM data will be downloaded.
Defaults to the countries in the Danube River Basin.}

\item{coords_df}{A data frame containing coordinates to spatially join with GADM administrative data.
It must include columns for latitude and longitude.}

\item{lat_col}{The name of the column in \code{coords_df} containing latitude values. Defaults to "decimalLatitude".}

\item{lon_col}{The name of the column in \code{coords_df} containing longitude values. Defaults to "decimalLongitude".}

\item{sf_gadm}{Optional \code{sf} object containing GADM administrative boundaries. If provided, it is used
directly for the spatial join. Defaults to \code{NULL}.}

\item{timeout}{Timeout in seconds for downloads. Defaults to 600 (10 minutes).}
}
\value{
A data frame with the same columns as the input \code{coords_df}, along with administrative attributes
from the GADM data.
}
\description{
This function downloads GADM administrative boundary data for specified countries
and performs a spatial join to add administrative information to a dataset containing
geographic coordinates. Optionally, an \code{sf} object can be provided instead of downloading data.
}
