% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_checklist}
\alias{species_checklist}
\title{Species list based on a preliminary fish species check list of the Danube
River Basin produced by the Join Danube Survey 5 (JDS5)}
\format{
\subsection{\code{species_checklist}}{

A character vector with species names. Length = 143
}
}
\source{
JDS5
}
\usage{
species_checklist
}
\description{
A character vector with species names
}
\keyword{datasets}
