% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_save_csv.R
\name{split_and_save_csv}
\alias{split_and_save_csv}
\title{Split a data frame based on the values of a column and save each subset as a CSV file}
\usage{
split_and_save_csv(df, split_col, output_dir = ".", sep = ",")
}
\arguments{
\item{df}{A data frame to be split.}

\item{split_col}{The name of the column to split the data frame by (as a string).}

\item{output_dir}{The directory where the CSV files will be saved. Default is the current working directory.}

\item{sep}{The separator used in the CSV files. Default is a comma (\verb{,}).}
}
\value{
No return value. CSV files are saved to disk.
}
\description{
This function splits a data frame by the unique values in a specified column and
saves each subset as a CSV file. The CSV files will be named based on the unique
values in the specified column.
}
\examples{
\dontrun{
# Assuming df has a column named "datasetName"
split_and_save_csv(df, split_col = "datasetName", output_dir = "output/")
}

}
