% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_points.R
\name{visualize_points}
\alias{visualize_points}
\title{Visualize Points on a Map with Optional Layers and Metadata}
\usage{
visualize_points(
  points_df,
  latitude_col = "decimalLatitude",
  longitude_col = "decimalLongitude",
  show_extra_columns = FALSE,
  layer = NULL
)
}
\arguments{
\item{points_df}{A data frame containing columns named 'decimalLatitude' and 'decimalLongitude'.
Optionally, the data frame can include other columns such as species, locality, year,
dataset name, or any additional metadata.}

\item{latitude_col}{The name of the latitude column in \code{points_df}. Default is 'decimalLatitude'.}

\item{longitude_col}{The name of the longitude column in \code{points_df}. Default is 'decimalLongitude'.}

\item{show_extra_columns}{A logical value indicating whether to display extra columns
(e.g., species, locality, year) in the popup. Default is FALSE.}

\item{layer}{An optional vector or raster layer (e.g., an sf object or a raster object)
to be added to the map. The points will be plotted on top of this layer.}
}
\value{
A \code{leaflet} map object with the points plotted on it, and optionally, an additional
vector or raster layer. The popup will display the coordinates and, if \code{show_extra_columns} is TRUE,
any additional columns provided.
}
\description{
This function takes a data frame containing latitude and longitude coordinates
(default columns: 'decimalLatitude' and 'decimalLongitude') and optionally additional
columns (e.g., species, locality, year, dataset name). It visualizes the points
on a map using the \code{leaflet} package, and the popup will show the coordinates along
with any other provided information if the user chooses to display it. Additionally,
it allows loading a vector or raster layer to the map, and the points will be displayed
on top of this layer.
}
\examples{
# Example: Visualize locations along the Danube River with fish species data
fish_data_2 <- fish_data \%>\%
select(species, decimalLatitude, decimalLongitude)
visualize_points(fish_data_2, show_extra_columns = TRUE)

# Example with additional vector layer
visualize_points(fish_data_2, layer = danube_basin, show_extra_columns = TRUE)

}
