% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_names.R
\name{check_column_names}
\alias{check_column_names}
\title{Check Column Names of a Data Frame}
\usage{
check_column_names(df, standard_names, verbose = FALSE)
}
\arguments{
\item{df}{A data frame to be checked.}

\item{standard_names}{A character vector containing the standard column names.}

\item{verbose}{Logical. If \code{TRUE}, prints the missing and extra columns.}
}
\value{
A list with two elements:
\itemize{
\item \code{missing_columns}: Column names that are in \code{standard_names} but missing from the data frame.
\item \code{extra_columns}: Column names that are in the data frame but not in \code{standard_names}.
}
}
\description{
This function checks if the column names of a data frame correspond to the
names contained in a vector of standard column names. It returns the columns
that are missing in the data frame or columns in the data frame that are
not part of the standard names vector. If \code{verbose} is \code{TRUE}, the function
prints details about the missing and extra columns.
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, d = 7:9)
standard_names <- c("a", "b", "c")
check_column_names(df, standard_names, verbose = TRUE)
}
