% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_subset.R
\name{get_spatial_subset}
\alias{get_spatial_subset}
\title{Spatial Subset of Coordinates within a Polygon}
\usage{
get_spatial_subset(
  polygon,
  data,
  lat_col,
  lon_col,
  crs = 4326,
  verbose = FALSE
)
}
\arguments{
\item{polygon}{An \code{sf} object representing the polygon for spatial filtering.}

\item{data}{A data frame containing at least two columns for latitude and longitude.}

\item{lat_col}{The name of the column in \code{data} representing latitude.}

\item{lon_col}{The name of the column in \code{data} representing longitude.}

\item{crs}{The CRS (Coordinate Reference System) to ensure consistency between the polygon and coordinates. Default is WGS84 (\code{EPSG:4326}).}

\item{verbose}{Logical. If \code{TRUE}, prints detailed messages. Default is \code{FALSE}.}
}
\value{
A data frame containing only the rows of the original data frame where
the coordinates fall within the polygon.
}
\description{
This function takes a polygon (\code{sf} object) and a data frame with latitude and
longitude columns, and filters the data frame to retain only the coordinates
that fall within the polygon. The CRS (Coordinate Reference System) of both
the polygon and the coordinates must match, and the default CRS is WGS84.
}
\examples{
# Example polygon and data
polygon <- danube_basin
data <- fish_data
filtered_data <- get_spatial_subset(polygon,
                                    data,
                                    "decimalLatitude",
                                    "decimalLongitude",
                                    verbose = TRUE)

# Map with points that fall within the polygon

}
