% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_name_app.R
\name{update_name_app}
\alias{update_name_app}
\title{Update Species Data in a Shiny MiniUI App}
\usage{
update_name_app(species_name)
}
\arguments{
\item{species_name}{A string containing species names.}
}
\value{
Returns a data frame with one column for the old names and another
for the new names.
}
\description{
This function launches a Shiny gadget within RStudio that allows users to
manually update species data. Users can select a species from a dropdown,
enter a new name, and update the corresponding entry. Once completed, the
updated names are saved and returned as a data frame with one column for
the old names and another for the new names.
}
\examples{
\dontrun{
# Try changing to correct names: Chondrostoma nasus, Hucho hucho,
# Anguilla anguilla
species_name <- c("Chondrostoma nasus", "Hucho hucho", "Anguilla anguilla")
updated_data <- update_name_app(species_name)
}

}
